+++ Runtime/UI/Styles.cs
@@ -133,5 +133,47 @@ namespace Unity.AppUI.UI
         /// The styling class used to set a "keyboard-focus" pseudo-state on a element.
         /// </summary>
         public const string keyboardFocusUssClassName = "keyboard-focus";
+        
+        /// <summary>
+        /// Converts a <see cref="Size"/> to a <see cref="IconSize"/>.
+        /// </summary>
+        /// <param name="size"> The size to convert. </param>
+        /// <returns> The converted <see cref="IconSize"/>. </returns>
+        /// <exception cref="ArgumentOutOfRangeException"> Thrown when an unknown <see cref="Size"/> is provided. </exception>
+        public static IconSize ToIconSize(this Size size)
+        {
+            switch (size)
+            {
+                case Size.S:
+                    return IconSize.S;
+                case Size.M:
+                    return IconSize.M;
+                case Size.L:
+                    return IconSize.L;
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(size), size, null);
+            }
+        }
+        
+        /// <summary>
+        /// Converts a <see cref="IconSize"/> to a <see cref="Size"/>.
+        /// </summary>
+        /// <param name="size"> The size to convert. </param>
+        /// <returns> The converted <see cref="Size"/>. </returns>
+        /// <exception cref="ArgumentOutOfRangeException"> Thrown when an unknown <see cref="IconSize"/> is provided. </exception>
+        public static Size ToSize(this IconSize size)
+        {
+            switch (size)
+            {
+                case IconSize.S:
+                    return Size.S;
+                case IconSize.M:
+                    return Size.M;
+                case IconSize.L:
+                    return Size.L;
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(size), size, null);
+            }
+        }
     }
 }